fix vbond bringup
The way the certificate is installed depends on the device
"personality". Vedges, vsmarts and vmanages have distinct
personalities, whereas vbond is just a running mode of a vedge.

The certificate installer script (vconfd_script_cert.sh) identifies a
vbond based on it's running configuration. The problem is the boothook
configuration today is applied only _after_ cloudinit runcmd is
done. So, the script installs the certificate as for a vedge into
/usr/share/viptela/vedge_certs/ as opposed to
/usr/share/viptela/server.crt.

After reboot, the newly applied configuration for vbond triggers the
launch of the vdaemon process which expects the server.crt in it's
usual location and crashes when it doesn't find it there.

The workaround is to simply copy the certificate to
/usr/share/viptela/server.crt.

This hack might be removed in the future when a proper
cloudinit boothook for Viptela configuration is implemented.