switch to 17.1 branch
The installed certificates are getting wiped out due to this commit
https://github.com/telsiz/viptela/commit/fc78ffba28b0706a04828902134cc7a299efb217
and the fact that the configuration is applied AFTER cloudinit runcmd.

Work-around it by setting the organization-name before installing the certs.

Also, vdaemon broke backwards compatibility after multi-tenancy changes.
Work-around it by adding the org-name to the generated serial number files.